requirements.txt
fastapi==0.104.0
uvicorn==0.24.0
python-dotenv==1.0.0
requests==2.31.0
chromadb==0.4.15
pydantic==2.4.2
beautifulsoup4==4.12.2
aiohttp==3.9.1